+++ src/value/reg_or.rs
@@ -10,7 +10,7 @@ pub enum RegOr<T : RegValue, U> {
 impl<T : RegValue, U : PacketEncode> PacketEncode for RegOr<T, U> { fn encode(&self, buf : &mut PacketBuf) -> Result<(), EncodeError> {
     match (self) {
         #[allow(deprecated)]
-        Self::Id(entry) => { buf.encode_write(VarInt::from(entry.id() + 1)) }
+        Self::Id(entry) => { buf.encode_write(VarInt::from((entry.id() as i32) + 1)) }
         Self::Or(value) => {
             buf.encode_write(VarInt::from(0))?;
             buf.encode_write(value)
@@ -19,7 +19,7 @@ impl<T : RegValue, U : PacketEncode> PacketEncode for RegOr<T, U> { fn encode(&s
 } }
 
 impl<T : RegValue, U : PacketDecode> PacketDecode for RegOr<T, U> { fn decode(buf : &mut PacketBuf) -> Result<Self, DecodeError> {
-    let id_plus_one = buf.read_decode::<VarInt>()?.as_i32() as usize;
+    let id_plus_one = buf.read_decode::<VarInt>()?.as_i32() as u32;
     Ok(if (id_plus_one == 0) {
         Self::Id(unsafe{ RegEntry::new_unchecked(id_plus_one - 1) })
     } else {
